// src/app/sistema/interfaces/variables.ts
export interface Laboratorio {
    id: string;
    descripcion: string;
    alias: string;

    datos: {};
    estado: number;

    createdBy: string;
    updatedBy: string;
    createdAt: string;
    updatedAt: string
}

export interface UnidadMedida {
    id: string;
    descripcion: string;
    alias: string;

    datos: {};
    estado: number;

    createdBy: string;
    updatedBy: string;
    createdAt: string;
    updatedAt: string
}

export interface Deposito {
    id: string;
    descripcion: string;
    alias: string;

    datos: {};
    estado: number;

    createdBy: string;
    updatedBy: string;
    createdAt: string;
    updatedAt: string
}